+++ packages/auth/src/SignUp/SignUpForm.tsx
@@ -4,20 +4,29 @@ import {
   FForm,
   FFormik,
   FInput,
+  FCheckbox,
   useT,
   validateLength,
   validateSchema,
   validateEmail,
+  validateCheckbox,
 } from '@frontegg/react-core';
 import { useAuthRoutes, useOnRedirectTo, useSignUpActions, useSignUpState } from '@frontegg/react-hooks/auth';
 import { FReCaptcha } from '../components/FReCaptcha';
+import { SignUpCheckbox } from './SignUp';
 const { Formik } = FFormik;
 
 export interface SignUpFormProps {
   withCompanyName?: boolean;
+  signUpConsent?: SignUpCheckbox;
+  marketingMaterialConsent?: SignUpCheckbox;
 }
 
-export const SignUpForm: FC<SignUpFormProps> = ({ withCompanyName = true }) => {
+export const SignUpForm: FC<SignUpFormProps> = ({
+  withCompanyName = true,
+  signUpConsent,
+  marketingMaterialConsent,
+}) => {
   const { t } = useT();
 
   const routes = useAuthRoutes();
@@ -25,9 +34,16 @@ export const SignUpForm: FC<SignUpFormProps> = ({ withCompanyName = true }) => {
   const { signUpUser } = useSignUpActions();
   const { loading } = useSignUpState(({ loading }) => ({ loading }));
 
-  const redirectToLogin = useCallback(() => {
-    onRedirectTo(routes.loginUrl);
+  const isCheckboxVisible = useCallback((checkbox?: SignUpCheckbox) => {
+    return checkbox?.hasOwnProperty('content');
   }, []);
+
+  const redirectToLogin = useCallback(() => {
+    onRedirectTo(routes.loginUrl, { preserveQueryParams: true });
+  }, []);
+
+  const isSignUpConsentVisible = isCheckboxVisible(signUpConsent);
+  const isMarketingMaterialVisible = isCheckboxVisible(marketingMaterialConsent);
 
   return (
     <>
@@ -44,18 +60,28 @@ export const SignUpForm: FC<SignUpFormProps> = ({ withCompanyName = true }) => {
           name: '',
           companyName: '',
           recaptchaToken: '',
+          acceptedTermsOfService: isSignUpConsentVisible ? false : undefined,
+          allowMarketingMaterial: isMarketingMaterialVisible ? false : undefined,
         }}
-        onSubmit={(values) => {
+        onSubmit={({ acceptedTermsOfService, allowMarketingMaterial, ...values }) => {
+          const metadata = JSON.stringify({
+            acceptedTermsOfService,
+            allowMarketingMaterial,
+          });
+
           if (withCompanyName) {
-            signUpUser(values);
+            signUpUser({ ...values, metadata });
           } else {
-            signUpUser({ ...values, companyName: values.name });
+            signUpUser({ ...values, companyName: values.name, metadata });
           }
         }}
         validationSchema={validateSchema({
           email: validateEmail(t),
           name: validateLength('name', 3, t),
           companyName: withCompanyName && validateLength('Company Name', 3, t),
+          acceptedTermsOfService: isSignUpConsentVisible && signUpConsent?.required !== false && validateCheckbox(),
+          allowMarketingMaterial:
+            isMarketingMaterialVisible && marketingMaterialConsent?.required !== false && validateCheckbox(),
         })}
       >
         <FForm>
@@ -75,7 +101,20 @@ export const SignUpForm: FC<SignUpFormProps> = ({ withCompanyName = true }) => {
               data-test-id='compenyName-box'
             />
           )}
-
+          {isSignUpConsentVisible && (
+            <FCheckbox
+              name='acceptedTermsOfService'
+              renderLabel={signUpConsent?.content}
+              className={'fe-sign-up__checkbox'}
+            />
+          )}
+          {isMarketingMaterialVisible && (
+            <FCheckbox
+              name='allowMarketingMaterial'
+              renderLabel={marketingMaterialConsent?.content}
+              className={'fe-sign-up__checkbox'}
+            />
+          )}
           <FButton type='submit' fullWidth variant='primary' loading={loading} data-test-id='signupSubmit-btn'>
             {t('auth.sign-up.form.submit-button')}
           </FButton>