+++ packages/auth/src/SocialLogins/types.ts
@@ -3,6 +3,7 @@ import { SocialLoginsProviders } from '@frontegg/rest-api';
 export interface ISocialLoginCallbackState {
   provider: SocialLoginsProviders;
   action: SocialLoginsActions;
+  afterAuthRedirectUrl?: string;
 }
 
 export interface ISocialLoginsContext {