+++ packages/auth/src/components/FReCaptcha.tsx
@@ -1,13 +1,30 @@
 import React, { FC, useEffect } from 'react';
 import { loadReCaptcha, ReCaptcha } from 'react-recaptcha-v3';
-import { useField } from 'formik';
 import { useCallback } from 'react';
+import { FFormik } from '@frontegg/react-core';
 import { useSecurityPolicyState } from '@frontegg/react-hooks/auth';
+
+const { useField } = FFormik;
 
 interface IReCaptchaProps {
   action: string;
   fieldName?: string;
 }
+
+const unload = (recaptchaSiteKey?: string) => {
+  if (!recaptchaSiteKey) {
+    return;
+  }
+  const nodeBadge = document.querySelector('.grecaptcha-badge');
+  if (nodeBadge && nodeBadge.parentNode) {
+    document.body.removeChild(nodeBadge.parentNode);
+  }
+  const scriptSelector = "script[src='https://www.google.com/recaptcha/api.js?render=" + recaptchaSiteKey + "']";
+  const script = document.querySelector(scriptSelector);
+  if (script) {
+    script.remove();
+  }
+};
 
 export const FReCaptcha: FC<IReCaptchaProps> = ({ action, fieldName = 'recaptchaToken' }) => {
   const { policy } = useSecurityPolicyState((state) => state.captchaPolicy);
@@ -15,7 +32,13 @@ export const FReCaptcha: FC<IReCaptchaProps> = ({ action, fieldName = 'recaptcha
 
   useEffect(() => {
     if (policy?.enabled) loadReCaptcha(policy?.siteKey || '');
+
+    return () => unload(policy?.siteKey);
   }, [policy?.enabled, policy?.siteKey]);
+
+  if (!policy?.enabled || !policy?.siteKey) {
+    return null;
+  }
 
   const handleCallback = useCallback(
     (token: string) => {
@@ -26,5 +49,5 @@ export const FReCaptcha: FC<IReCaptchaProps> = ({ action, fieldName = 'recaptcha
 
   if (!policy || !policy.enabled) return null;
 
-  return <ReCaptcha sitekey={policy.siteKey} verifyCallback={handleCallback} action={action} />;
+  return <ReCaptcha sitekey={policy?.siteKey} verifyCallback={handleCallback} action={action} />;
 };