+++ packages/core/src/elements/Checkbox/interfaces.ts
@@ -3,6 +3,7 @@ import { FormFieldProps } from '../../ElementsFactory';
 
 export interface CheckboxProps extends Omit<InputHTMLAttributes<HTMLInputElement>, 'value' | 'size'>, FormFieldProps {
   label?: string;
+  renderLabel?: () => JSX.Element;
   indeterminate?: boolean;
   fullWidth?: boolean;
   error?: string;