+++ packages/redux-store/src/auth/LoginState/saga.ts
@@ -20,7 +20,10 @@ export function* afterAuthNavigation() {
   window.localStorage.removeItem(FRONTEGG_AFTER_AUTH_REDIRECT_URL);
   yield delay(200);
   put(actions.resetLoginState());
-  onRedirectTo(authenticatedUrl);
+  const url = new URL(window?.location.href);
+  const redirectUrl = url.searchParams.get('redirectUrl') || authenticatedUrl;
+
+  onRedirectTo(redirectUrl, { refresh: redirectUrl.startsWith('http') });
 }
 
 function* refreshMetadata() {
@@ -79,7 +82,7 @@ export function* refreshToken() {
           },
         })
       );
-      onRedirectTo(routes.loginUrl);
+      onRedirectTo(routes.loginUrl, { preserveQueryParams: true });
     } else {
       ContextHolder.setAccessToken(user.accessToken);
       ContextHolder.setUser(user);
@@ -251,9 +254,7 @@ function* logout({ payload }: PayloadAction<() => void>) {
   yield put(actions.setState({ isLoading: true }));
   try {
     yield call(api.auth.logout);
-  } catch (e) {
-    console.error(e);
-  }
+  } catch {}
   yield put(actions.resetState());
   yield put(actions.requestAuthorize(true));
   payload?.();
@@ -262,9 +263,7 @@ function* logout({ payload }: PayloadAction<() => void>) {
 function* silentLogout({ payload }: PayloadAction<() => void>) {
   try {
     yield call(api.auth.logout);
-  } catch (e) {
-    console.error(e);
-  }
+  } catch {}
   payload?.();
 }